src/pyproject.toml
[tool.poetry]
package-mode = false
name = "backend-login-app"
version = "0.1.0"
description = ""
authors = ["Mudit Gurjar <mudit182@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
sec = "^0.3.0"
sqlalchemy-utils = "^0.41.1"
sqlalchemy = "^2.0.28"
redis = "^5.0.3"
pymongo = "^4.6.2"
pillow = "^10.2.0"
requests = "^2.31.0"
aiofiles = "^23.2.1"
aiostream = "^0.5.2"
typing-extensions = "^4.10.0"
pydantic = "^2.6.4"
pymysql = "^1.1.0"
bcrypt = "^4.1.2"
pydantic-settings = "^2.2.1"
pydantic-mongo = "^2.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-timeout = "^2.3.1"
asgi-lifespan = "^2.1.0"
pytest-xdist = "^3.5.0"
pytest-watch = "^4.2.0"
pytest-html = "^4.1.1"
pytest-testmon = "^2.1.1"
allure-pytest = "^2.13.3"
coverage = "^7.4.4"
mimesis = "^15.1.0"
pre-commit = "^3.6.2"
pymongo-inmemory = "^0.4.1"
pytest-playwright = "^0.4.4"
httpx = "0.26"
httpx-ws = "0.4.3"
pylint = "^3.1.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
